REFACTOR: Logging
-Runtime log now tracks the keyword

-Added OutputController to ensure that the QueryController was only responsible for queries
