+++ Logs/runtime.csv
@@ -1,8 +1,20 @@
-MySQL,266.0
-MongoDB,483.0
-Lucene Index,4317.0
-Brute Force,17546.0
-MySQL,185.0
-MongoDB,469.0
-Lucene Index,837.0
-Brute Force,14639.0
+MySQL,flu,146.0
+MongoDB,flu,276.0
+Lucene Index,flu,1264.0
+Brute Force,flu,11203.0
+MySQL,covid-19,169.0
+MongoDB,covid-19,187.0
+Lucene Index,covid-19,2240.0
+Brute Force,covid-19,11242.0
+MySQL,stomach,150.0
+MongoDB,stomach,211.0
+Lucene Index,stomach,505.0
+Brute Force,stomach,11034.0
+MongoDB,stomach,187.0
+MySQL,stomach,232.0
+Lucene Index,stomach,1351.0
+Brute Force,stomach,11990.0
+MongoDB,brain,203.0
+MySQL,brain,228.0
+Lucene Index,brain,806.0
+Brute Force,brain,12077.0
